pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "misst"
version = "0.1.2"
description = "Automated murine polysomnogram sleep staging for Python."
readme = "docs/README_PYPI.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [
    {name = "Hudson Liu", email = "hudsonliu0@gmail.com"},
]
keywords = ["misst", "sleep staging", "murine", "polysomnogram", "machine learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License"
]
dependencies = [
    "keras >= 2.9.0",
    "keras_tuner >= 1.1.3",
    "matplotlib >= 3.6.2",
    "mne >= 1.3.0",
    "numpy >= 1.22.3",
    "overrides >= 7.3.1",
    "pandas >= 1.5.2",
    "pyyaml >= 6.0.1",
    "scikit_learn >= 1.3.0",
    "tensorflow >= 2.9.1",
    "tensorflow_addons >= 0.20.0",
    "tqdm >= 4.64.1",
    "EDFlib-Python >= 1.0.7"
]

[project.urls]
Repository = "https://github.com/Johns-Hopkins-CISRE/MISST"

[project.optional-dependencies]
dev = ["twine>=4.0.2"]
